+++ R/focep.R
@@ -116,6 +116,7 @@ nlmixr2Est.focep <- function(env, ...) {
   .ui <- env$ui
   .foceiFamilyReturn(env, .ui, ..., est="focep")
 }
+attr(nlmixr2Est.focep, "nlmixr2Priors") <- "general"
 attr(nlmixr2Est.focep, "iov") <- TRUE
 attr(nlmixr2Est.focep, "covPresent") <- TRUE
 attr(nlmixr2Est.focep, "unbounded") <- .foUnbounded