+++ R/ifocep.R
@@ -114,6 +114,7 @@ nlmixr2Est.ifocep <- function(env, ...) {
   .ui <- env$ui
   .foceiFamilyReturn(env, .ui, ..., est="ifocep")
 }
+attr(nlmixr2Est.ifocep, "nlmixr2Priors") <- "general"
 attr(nlmixr2Est.ifocep, "iov") <- TRUE
 attr(nlmixr2Est.ifocep, "covPresent") <- TRUE
 attr(nlmixr2Est.ifocep, "unbounded") <- .foUnbounded